reuse setRoutes
